\name{stnlist}
\Rdversion{1.1}
\alias{stnlist}
\docType{data}
\title{
CIMIS station list
}
\description{
List of CIMIS station numbers and locations.
}
\usage{data(stnlist)}
\format{
  A data frame with 133 observations on the following 2 variables.
  \describe{
    \item{\code{station}}{station number}
    \item{\code{name}}{Station name}
  }
}
\source{
\url{http://www.cimis.water.ca.gov/cimis/frontStationListData.do}
}

\examples{
data(stnlist)

}
\keyword{datasets}
